requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blockbuster==1.5.24
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
feedparser==6.0.11
forbiddenfruit==0.1.4
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-rs==0.29.1
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.17
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-api==0.2.34
langgraph-checkpoint==2.0.26
langgraph-cli==0.2.10
langgraph-prebuilt==0.2.1
langgraph-runtime-inmem==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
marshmallow==3.26.1
multidict==6.4.4
mypy==1.15.0
mypy-extensions==1.1.0
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pip==25.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyjwt==2.10.1
pypdf==5.5.0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
